Adicionando reference ao EntityFramework 6.2 e criando a estrutura de Infra

+++ GerenciadorPedidos.Domain/Argumentos/Produtos/ProdutoAlterarRequest.cs
@@ -7,7 +7,7 @@ namespace GerenciadorPedidos.Domain.Argumentos.Produtos
         public Guid id { get; set; }
         public string descricao { get; set; }
         public Guid categoriaProdutoId { get; set; }
-        public double valorVenda { get; set; }
+        public decimal valorVenda { get; set; }
         public int quantidade { get; set; }
     }
 }
